requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-dialogflow==2.10.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
grpcio-status==1.42.0
idna==3.3
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-decouple==3.5
python-telegram-bot==13.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
rsa==4.8
six==1.16.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
